+++ pigit/termui/__init__.py
@@ -8,10 +8,10 @@ Date: 2026-03-26
 
 from __future__ import annotations
 
-from pigit.termui.component import Component
+from pigit.termui.components import Component
 from pigit.termui.geometry import TerminalSize
 from pigit.termui.input_keyboard import KeyboardInput
-from pigit.termui.loop import run_key_echo
+from pigit.termui.key_echo import run_key_echo
 from pigit.termui.render import Renderer
 from pigit.termui.session import Session
 from pigit.termui.text import get_width, plain